admin_ui/src/requirements.txt
aioredis==2.0.0
aiosmtplib==1.1.6
anyio==3.3.2
asgiref==3.4.1
astroid==2.7.3
async-timeout==3.0.1
attrs==21.2.0
backports.entry-points-selectable==1.1.0
black==21.8b0
blinker==1.4
certifi==2021.5.30
cfgv==3.3.1
charset-normalizer==2.0.6
click==8.0.1
colorama==0.4.4
commonmark==0.9.1
distlib==0.3.2
dnspython==2.1.0
email-validator==1.1.3
fakeredis==1.6.1
fastapi==0.68.1
fastapi-mail==1.0.1
filelock==3.0.12
greenlet==1.1.1
h11==0.12.0
httpcore==0.13.7
httpx==0.19.0
identify==2.2.13
idna==3.2
iniconfig==1.1.1
isort==5.9.3
Jinja2==3.0.1
lazy-object-proxy==1.6.0
MarkupSafe==2.0.1
mccabe==0.6.1
mypy==0.910
mypy-extensions==0.4.3
nodeenv==1.6.0
packaging==21.0
pathspec==0.9.0
platformdirs==2.3.0
pluggy==1.0.0
pprintpp==0.4.0
pre-commit==2.15.0
prometheus-client==0.11.0
py==1.10.0
pydantic==1.8.2
Pygments==2.10.0
pylint==2.10.2
pyparsing==2.4.7
pytest==6.2.5
pytest-clarity==1.0.1
pytest-ordering==0.6
python-dotenv==0.19.0
python-multipart==0.0.5
PyYAML==5.4.1
redis==3.5.3
regex==2021.8.28
requests==2.26.0
rfc3986==1.5.0
rich==10.9.0
six==1.16.0
sniffio==1.2.0
sortedcontainers==2.4.0
SQLAlchemy==1.4.23
starlette==0.14.2
starlette-exporter==0.10.0
toml==0.10.2
tomli==1.2.1
typing-extensions==3.10.0.2
urllib3==1.26.7
uvicorn==0.15.0
virtualenv==20.7.2
wrapt==1.12.1
